+++ api/src/main/java/org/allaymc/api/world/WorldViewer.java
@@ -147,8 +147,6 @@ public interface WorldViewer {
      */
     void viewChunk(Chunk chunk);
 
-    // TODO: viewSubChunk()
-
     /**
      * Removes the chunk passed.
      *