+++ server/src/main/java/org/allaymc/server/network/NetworkData.java
@@ -8,9 +8,7 @@ import org.allaymc.api.pack.PackManifest;
 import org.allaymc.api.registry.Registries;
 import org.allaymc.api.server.Server;
 import org.allaymc.api.utils.Utils;
-import org.allaymc.api.world.biome.BiomeId;
 import org.allaymc.server.registry.InternalRegistries;
-import org.allaymc.server.world.biome.BiomeData;
 import org.cloudburstmc.nbt.NbtMap;
 import org.cloudburstmc.nbt.NbtUtils;
 import org.cloudburstmc.protocol.bedrock.data.ExperimentData;
@@ -104,9 +102,8 @@ public final class NetworkData {
 
     public static BiomeDefinitionListPacket encodeBiomeDefinitionListPacket() {
         Map<String, BiomeDefinitionData> definitions = new HashMap<>();
-        for (var biomeId : BiomeId.values()) {
-            var biomeData = BiomeData.getBiomeData(biomeId);
-            definitions.put(biomeId.getIdentifier().path(), biomeData.toNetworkData());
+        for (var biome : Registries.BIOMES.getContent().m1().values()) {
+            definitions.put(biome.getIdentifier().path(), NetworkHelper.toNetwork(biome));
         }
         var packet = new BiomeDefinitionListPacket();
         packet.setBiomes(new BiomeDefinitions(definitions));