+++ server/src/main/java/org/allaymc/server/registry/InternalRegistries.java
@@ -5,10 +5,12 @@ import org.allaymc.api.block.property.type.BlockPropertyType;
 import org.allaymc.api.item.data.ItemData;
 import org.allaymc.api.registry.SimpleMappedRegistry;
 import org.allaymc.api.registry.SimpleRegistry;
+import org.allaymc.api.world.biome.BiomeData;
 import org.allaymc.server.block.data.BlockId;
 import org.allaymc.server.block.property.BlockPropertyProcessor;
 import org.allaymc.server.item.data.ItemId;
 import org.allaymc.server.item.type.ItemComponentData;
+import org.allaymc.server.world.biome.BiomeId;
 import org.cloudburstmc.protocol.bedrock.data.TrimMaterial;
 import org.cloudburstmc.protocol.bedrock.data.TrimPattern;
 
@@ -25,6 +27,8 @@ public final class InternalRegistries {
     public static SimpleMappedRegistry<BlockPropertyType<?>, BlockPropertyProcessor> BLOCK_PROPERTY_PROCESSORS;
 
     public static SimpleMappedRegistry<ItemId, ItemComponentData> ITEM_COMPONENT_DATA;
+
+    public static SimpleMappedRegistry<BiomeId, BiomeData> BIOME_DATA;
 
     public static SimpleRegistry<Set<TrimPattern>> TRIM_PATTERNS;
     public static SimpleRegistry<Set<TrimMaterial>> TRIM_MATERIALS;