+++ server/src/test/java/org/allaymc/server/registry/RegistryTest.java
@@ -1,9 +1,9 @@
-package org.allaymc.server.block.registry;
+package org.allaymc.server.registry;
 
 import org.allaymc.api.registry.Registries;
 import org.allaymc.server.block.data.BlockId;
 import org.allaymc.server.item.data.ItemId;
-import org.allaymc.server.registry.InternalRegistries;
+import org.allaymc.server.world.biome.BiomeId;
 import org.allaymc.testutils.AllayTestExtension;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -23,13 +23,16 @@ public class RegistryTest {
     }
 
     @Test
-    void testBlockAttributeRegistry() {
+    void testBlockStateDataRegistry() {
+        for (var id : BlockId.values()) {
+            assertNotNull(InternalRegistries.BLOCK_STATE_DATA.get(id));
+        }
         //Special case
         assertNotNull(InternalRegistries.BLOCK_STATE_DATA.get(BlockId.UNKNOWN).get(-2));
     }
 
     @Test
-    void testItemAttributeRegistry() {
+    void testItemDataRegistry() {
         for (var id : ItemId.values()) {
             assertNotNull(InternalRegistries.ITEM_DATA.get(id));
         }
@@ -39,6 +42,20 @@ public class RegistryTest {
     void testItemTypeRegistry() {
         for (var id : ItemId.values()) {
             assertNotNull(Registries.ITEMS.get(id.getIdentifier()));
+        }
+    }
+
+    @Test
+    void testBiomeDataRegistry() {
+        for (var id : BiomeId.values()) {
+            assertNotNull(InternalRegistries.BIOME_DATA.get(id));
+        }
+    }
+
+    @Test
+    void testBiomeRegistry() {
+        for (var id : BiomeId.values()) {
+            assertNotNull(Registries.BIOMES.getByK2(id.getIdentifier()));
         }
     }
 }